Fix a problem with FloodFill
The logs added for the score at which matching stopped in findAll,
helped to find and fix a case with floodFill where the whole result Mat got filled to zero. Solved it by using FLOODFILL_FIXED_RANGE flag (identified the problem by seeing that matching stopped at a score of zero).